refactor: move settings actions to one file
